+++ lib/src/models/seatsio_serializers.g.dart
@@ -21,15 +21,6 @@ Serializers _$serializers = (new Serializers().toBuilder()
       ..add(TicketListing.serializer)
       ..add(TicketTypePricing.serializer)
       ..addBuilderFactory(
-          const FullType(BuiltList, const [const FullType(String)]),
-          () => new ListBuilder<String>())
-      ..addBuilderFactory(
-          const FullType(BuiltList, const [const FullType(TicketTypePricing)]),
-          () => new ListBuilder<TicketTypePricing>())
-      ..addBuilderFactory(
-          const FullType(BuiltList, const [const FullType.nullable(String)]),
-          () => new ListBuilder<String?>())
-      ..addBuilderFactory(
           const FullType(BuiltList, const [const FullType(PricingForCategory)]),
           () => new ListBuilder<PricingForCategory>())
       ..addBuilderFactory(
@@ -58,7 +49,13 @@ Serializers _$serializers = (new Serializers().toBuilder()
           () => new MapBuilder<String, String>())
       ..addBuilderFactory(
           const FullType(BuiltList, const [const FullType(String)]),
-          () => new ListBuilder<String>()))
+          () => new ListBuilder<String>())
+      ..addBuilderFactory(
+          const FullType(BuiltList, const [const FullType(String)]),
+          () => new ListBuilder<String>())
+      ..addBuilderFactory(
+          const FullType(BuiltList, const [const FullType(TicketTypePricing)]),
+          () => new ListBuilder<TicketTypePricing>()))
     .build();
 
 // ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,deprecated_member_use_from_same_package,lines_longer_than_80_chars,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new