+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.0.2
+* [Feature] Update some properties of SeatingChartConfig.
+* [Feature] Update example.
+* [Feature] Update readme document.
 
-* Create Seatsio SDK for Flutter.
+## 0.0.1
+* [Feature] Create Seatsio SDK for Flutter.